drm/msm: Reduce latency while completing non-blocking commits
In order to complete a commit, we must first wait for the previous
commit to finish up. This is done by sleeping, during which time the CPU
can enter a deep idle state and take a while to finish processing the
commit after the wait is over. We can alleviate this by optimistically
assuming that the kthread this commit worker is running on won't migrate
to another CPU partway through. We only do this for the non-blocking
case where the commit completion is done in an asynchronous worker
because the generic DRM code already does this for atomic ioctls. Since
non-blocking commits are the common case, preemptively kick the prime
CPU out of idle when the atomic ioctl runs so that it'll be ready to
process the DRM IRQ and display commit kthreads.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[@0ctobot: Adapt for 4.19]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: LibXZR <xzr467706992@163.com>
Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>